% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDCasMat.R
\name{extractDCasMat}
\alias{extractDCasMat}
\title{Calculate dipeptide frequency matrix}
\usage{
extractDCasMat(seqvec, count = TRUE)
}
\arguments{
\item{seqvec}{A character vector of peptide/protein sequences.}

\item{count}{Logical, default TRUE. If TRUE, returns total count of dipeptide. If FALSE, returns frequency.}
}
\value{
A matrix of dipeptide frequency/count, with its row being the first amino acid and its column being the second amino acid.
}
\description{
This function calculates a dipeptide frequency (or count) matrix from a vector of sequences.
}
\examples{
\dontrun{
# An artificial, very short example
seq=c("AHGTYIR", "VCSDWEQLPMK", "FGGNTIPCAWMR", "QTYILMNK", "ASTEMNDNK")
dipepmat=extractDCasMat(seq)
print(dipepmat) # Note that the presence of two "NK" motifs
lattice::levelplot(dipepmat, xlab="First", ylab="Second", col.regions=heat.colors)
}
}

