% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCleavage.R
\name{plotCleavage}
\alias{plotCleavage}
\title{Plot cleavage selectivity as a barchart from identified peptide sequences}
\usage{
plotCleavage(seq, after, hide.plot = FALSE, normalize = FALSE)
}
\arguments{
\item{seq}{A character vector of identified peptide sequences without any modifications. Should only made up of 20 amino acids one letter characters.}

\item{after}{A character vector of peptide sequences after cleavage site. Only the first amino acid is used, but can be of longer length.}

\item{hide.plot}{Logical, default FALSE. If set to TRUE, no plot will be drawn. Can be used just to obtain a data.frame of cleavage site frequency for manual plotting.}

\item{normalize}{Logical, default FALSE (stacked barchart of cleaved and not-cleaved).}
}
\value{
If (and only if) assigned to a variable, this function returns a data.frame with columns (\code{AA}, \code{Freq}, \code{terminal}, \code{cleavage}) for custom plotting.
}
\description{
This function allows to examine specificity of enzymatic digestion (or chemical cleavage), based on MS/MS identification result (search without any enzyme specification, or "unspecific" search). Can be useful to check cleavage selectivity & efficiency of each enzyme used in multiple digestion or of less-specific enzymes.
}
\examples{
data(Elastase)
plotCleavage(Elastase$Sequence, Elastase$AAs.After)
plotCleavage(Elastase$Sequence, Elastase$AAs.After, normalize=TRUE)
}
\seealso{
\code{\link{plotCleavageMatrix}}
}

