% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCleavageMatrix.R
\name{plotCleavageMatrix}
\alias{plotCleavageMatrix}
\title{Plot cleavage site matrix from identified peptide sequences}
\usage{
plotCleavageMatrix(seq, after, hide.plot = FALSE, normalize = FALSE,
  col.regions = NULL)
}
\arguments{
\item{seq}{A character vector of identified peptide sequences without any modifications. Should only made up of 20 amino acids one letter characters.}

\item{after}{A character vector of peptide sequences after cleavage site. Only the first amino acid is used, but can be of longer length.}

\item{hide.plot}{Logical, default FALSE. If set to TRUE, no plot will be drawn. Can be used just to obtain a data.frame of cleavage site frequency for manual plotting.}

\item{normalize}{Logical, default FALSE (stacked barchart of cleaved and not-cleaved).}

\item{col.regions}{Color palette function like \code{heat.colors} etc.}
}
\value{
If assigned to a variable, this function returns a cleavage site matrix, with its row being the first amino acid and its column being the second amino acid.
}
\description{
This function allows to examine specificity of enzymatic digestion (or chemical cleavage) at dipeptide sequence level, based on MS/MS identification result (search without any enzyme specification, or "unspecific" search). Can be useful to check cleavage selectivity & efficiency of each enzyme used in multiple digestion or of less-specific enzymes.
}
\examples{
data(LysCTryp)
plotCleavageMatrix(LysCTryp$Sequence, LysCTryp$AAs.After, normalize=TRUE)
}
\seealso{
\code{\link{plotCleavage}}, \code{\link{extractDCasMat}}
}

